Add files via upload
v9 doesn't have break_out_of_loop function that makes sure that the back_up function doesn't continually execute as the robot backs up again and again around the circle, eventually falling out of the circle. 

v10 does have this function. Only issue now is that I don't know how to prevent the run_away function from running the robot out of the circle if we're near the edge. There should be an override while it crosses the black line in the run_away function (defense position), but it doesn't seem to work.